requirements.txt
fastapi
uvicorn
langchain
langserve
python-dotenv
